Move all the parsing models to the parser module.
